Move Client to deltakit_core
