+++ deltakit-core/pyproject.toml
@@ -34,6 +34,10 @@ dependencies = [
   "stim>=1.12.1, <2.0",
   "typing_extensions>=4.0, <5.0",
   "tqdm >=4.66.2",
+  "requests >=2.32.0",
+  "gql >=3.5.0",
+  "graphql_core >=3.2.0",
+  "requests_toolbelt >=1.0.0",  # needed by gql, which doesn't automatically install it
 ]
 
 [project.urls]
