+++ deltakit-core/src/deltakit_core/api/client/__init__.py
@@ -1,0 +1,3 @@
+from ._client import Client
+
+__all__ = ["Client"]
