+++ deltakit-core/src/deltakit_core/api/client/_api_client.py
@@ -9,9 +9,9 @@ import numpy.typing as npt
 import stim
 from deltakit_core.api.enums import APIEndpoints
 from deltakit_core.types import (Decoder, DecodingResult, DetectionEvents,
-                                     LeakageFlags, Measurements, NoiseModel,
-                                     ObservableFlips,
-                                     QubitCoordinateToDetectorMapping)
+                                 LeakageFlags, Measurements, NoiseModel,
+                                 ObservableFlips,
+                                 QubitCoordinateToDetectorMapping)
 from deltakit_core.types._experiment_types import QECExperimentDefinition
