+++ deltakit-explorer/pyproject.toml
@@ -29,15 +29,11 @@ classifiers = [
   "Typing :: Typed",
 ]
 dependencies = [
-    "gql >=3.5.0",
-    "graphql_core >=3.2.0",
     "numpy >=1.26",
     "stim >=1.9.0",
     "matplotlib >=3.5.2",
-    "requests >=2.32.0",
     "seaborn >=0.12.0",
     "typing-extensions >=4.0.0",
-    "requests_toolbelt >=1.0.0",  # needed by gql, which doesn't automatically install it
     "pymatching >= 2.2.2",
     "galois >= 0.4.6",
     "bposd >= 2.1",
